requirements.txt
flask
requests_oauthlib
requests
discord.py